refac: mv create patch extractor funcs to factory file; remove only constructor protocol
