+++ src/careamics/dataset_ng/patch_extractor/__init__.py
@@ -1,7 +1,6 @@
 __all__ = [
     "PatchExtractor",
-    "PatchExtractorConstructor",
     "PatchSpecs",
 ]
 
-from .patch_extractor import PatchExtractor, PatchExtractorConstructor, PatchSpecs
+from .patch_extractor import PatchExtractor, PatchSpecs
